Added Dockerfile
Gevolution can now be built and run with Docker,
which can be useful for just running the code
without the need to install any additional packages
or libraries on the machine (except Docker itself).

The current version builds gevolution with
CLASS and HEALPix support, as well as the lccat
and lcmap utilities.